1. fix driver download file broken, 2. move annotation define to api package
