+++ api/consts/annotations.go
@@ -1,0 +1,3 @@
+package consts
+
+const PodAnnotationSMCR = "network.alibabacloud.com/erdma-smcr"
