+++ internal/consts/consts.go
@@ -3,7 +3,5 @@ package consts
 const (
 	UA = "alibabacloud-erdma-controller"
 
-	AnnotationSMCR = "network.alibabacloud.com/erdma-smcr"
-
 	SMCRPNETEnv = "SMCR_PNET"
 )
